Example of QA tests that include Service Events
User service QA tests were added start of Embedded Event Broker.

New test feature was created with simple scenario that uses Embedded Event
Broker. This scenario creates account and user and then deletes account and
expects user to be deleted to.
It fails!
Service event is not working in this case, as when account is deleted, users
in that account should be deleted too.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>